% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PicoP2PFutureBackend-class.R
\name{availablePicoP2PWorkers}
\alias{availablePicoP2PWorkers}
\title{Gets the number of known P2P workers}
\usage{
availablePicoP2PWorkers()
}
\value{
\code{availablePicoP2PWorkers()} returns the number of registered workers on the P2P cluster.
It will always return at least one worker, which is yourself.
\emph{WARNING: This is currently hardcoded to 10 workers, regardless of the number.}
}
\description{
Gets the number of known P2P workers
}
\keyword{internal}
