% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoogleDriveP2PFutureBackend-class.R
\name{google_drive_p2p}
\alias{google_drive_p2p}
\alias{availableGoogleDriveP2PWorkers}
\title{google_drive_p2p futures}
\usage{
google_drive_p2p(
  ...,
  workers = availableGoogleDriveP2PWorkers(),
  envir = parent.frame()
)

availableGoogleDriveP2PWorkers()
}
\arguments{
\item{workers}{The number of concurrent workers to use.}

\item{envir}{The \link{environment} from where global objects should be
identified.}

\item{\ldots}{Additional arguments passed to \code{Future()}.}
}
\value{
An object of class \code{GoogleDriveP2PFuture}.

\code{availableGoogleDriveP2PWorkers()} returns the number of registered workers on the P2P cluster.
It will always return at least one worker, which is yourself.
\emph{WARNING: This is currently hardcoded to 10 workers, regardless of the number.}
}
\description{
\emph{WARNING: This function must never be called.
It may only be used with \code{\link[future:plan]{future::plan()}}}
}
\details{
A google_drive_p2p future is an asynchronous multiprocess
future that will be evaluated in a background R session.
}
