% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoogleDriveP2PFutureWorkers-class.R
\name{google_drive_p2p_worker}
\alias{google_drive_p2p_worker}
\title{Launches a Google Drive P2P Worker}
\usage{
google_drive_p2p_worker()
}
\description{
Launches a Google Drive P2P Worker
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
googledrive::drive_auth()
google_drive_p2p_worker()
\dontshow{\}) # examplesIf}
}
