% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoogleDriveP2PFutureBackend-class.R
\name{GoogleDriveP2PFutureBackend}
\alias{GoogleDriveP2PFutureBackend}
\title{A Google Drive P2P future is resolved through a Peer-to-Peer (P2P) workers communicating via Google Drive}
\usage{
GoogleDriveP2PFutureBackend(workers = availableGoogleDriveP2PWorkers(), ...)
}
\arguments{
\item{workers}{(optional) The maximum number of workers the
backend may use at any time.}

\item{\ldots}{Additional arguments passed to \code{\link[future:FutureBackend]{future::FutureBackend()}}.}
}
\value{
A GoogleDriveP2PFutureBackend object
}
\description{
A Google Drive P2P future is resolved through a Peer-to-Peer (P2P) workers communicating via Google Drive
}
\keyword{internal}
