% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2p_client_id.R
\name{p2p_client_id}
\alias{p2p_client_id}
\alias{p2p_worker_id}
\title{Gets the identifier of the current P2P client or P2P worker}
\usage{
p2p_client_id(...)

p2p_worker_id(...)
}
\value{
\code{p2p_client_id()} and \code{p2p_worker_id()} return the client and worker
identifier, which both have format \code{"{username}@{hostname}:{pid}"}.
}
\description{
Gets the identifier of the current P2P client or P2P worker
}
\keyword{internal}
