% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2p_client_id.R
\name{p2p_cluster_name}
\alias{p2p_cluster_name}
\title{Gets the default name of the P2P cluster}
\usage{
p2p_cluster_name(users = character(0))
}
\arguments{
\item{users}{Users to have access to the cluster. This controls whether
the default cluster names should be "personal" or "friends".}
}
\value{
\code{p2p_cluster_name()} returns R option \code{future.p2p.cluster}, if set.
If not set, it returns \code{{pico_name}/personal} if \code{length(users) == 0},
otherwise \code{{pico_name}/friends}.
}
\description{
Gets the default name of the P2P cluster
}
