% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2p_name.R
\name{p2p_name}
\alias{p2p_name}
\title{Gets the default name of the client or worker on the P2P cluster}
\usage{
p2p_name()
}
\value{
Returns \verb{\{username\}@\{hostname\}:\{pid\}}.
}
\description{
Gets the default name of the client or worker on the P2P cluster
}
