% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pico_p2p_cluster.R
\name{pico_p2p_cluster}
\alias{pico_p2p_cluster}
\title{Launches a P2P cluster}
\usage{
pico_p2p_cluster(
  cluster = "mycluster",
  users = character(0L),
  name = p2p_name(),
  host = "pipe.pico.sh",
  ssh_args = NULL,
  duration = 14 * 24 * 60 * 60
)
}
\arguments{
\item{cluster}{(character string) The name of the p2p cluster.}

\item{users}{(character vector) Names of Pico users who should have access,
in addition to the owner.}

\item{host}{(character string) The hostname serving the pico service.}

\item{ssh_args}{(character vector) Optional SSH arguments.}

\item{duration}{Duration (in seconds) to offer this cluster.}
}
\description{
Launches a P2P cluster
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pico_p2p_cluster()
\dontshow{\}) # examplesIf}
}
