% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pico.R
\name{pico_pipe}
\alias{pico_pipe}
\alias{pico_terminate}
\alias{pico_send_message}
\alias{pico_send_message_dataframe}
\alias{pico_receive_message}
\alias{pico_receive_message_dataframe}
\alias{pico_hosted_channels}
\title{Connect to a pico pipe}
\usage{
pico_pipe(
  topic = NULL,
  command = c("pipe", "pub", "sub", "ls", "help"),
  args = c(),
  host = "pipe.pico.sh",
  ssh_args = NULL,
  ...
)

pico_terminate(p, ...)

pico_send_message(p, message, newline = TRUE, ...)

pico_send_message_dataframe(p, df)

pico_receive_message(p, n = 1L, ...)

pico_receive_message_dataframe(p, ..., pattern = NULL)

pico_hosted_channels(host = "pipe.pico.sh", ssh_args = NULL, timeout = 10)
}
\arguments{
\item{topic}{(character string) The topic to connect to. If connecting
to another user's topic, prefix with their username and a forward slash,
e.g. \code{alice/topic}.}

\item{command}{(character string) Type of pipe or command.}

\item{args}{(character vector) Optional arguments.}

\item{host}{(character string) The hostname serving the pico service.}

\item{ssh_args}{(character vector) Optional SSH arguments.}

\item{p}{A \code{pico_pipe} object.}

\item{message}{(character string) A message to send.}

\item{newline}{(logical) If TRUE, a newline (LF; \verb{\\n}) will be appended
to the message, otherwise not.}

\item{df}{(data.frame) Data frame to send as a message.}

\item{n}{(integer) Number of messages to read.}

\item{pattern}{(character string; optional) A regular expression to scan for.}

\item{timeout}{(numeric scalar) Timeout (in seconds).}

\item{\ldots}{(character vector; optional) Named attributes that are recorded
with the returned object.}
}
\value{
A \link[processx:process]{processx::process} of class \code{pico_pipe}.
}
\description{
Connect to a pico pipe
}
\seealso{
This function relies on the \url{https://pico.sh} services.
}
\keyword{internal}
