% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pico_p2p.R
\name{pico_username}
\alias{pico_username}
\title{Gets your pico.sh username}
\usage{
pico_username(host = "pipe.pico.sh", ssh_args = NULL, timeout = 10)
}
\arguments{
\item{host}{(character string) The hostname serving the pico service.}

\item{ssh_args}{(character vector) Optional SSH arguments.}

\item{timeout}{(numeric scalar) Maximum number of seconds before giving up.}
}
\value{
A character string or a timeout error.
}
\description{
Gets your pico.sh username
}
