% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-options.R
\name{zzz-future.p2p.options}
\alias{zzz-future.p2p.options}
\alias{future.p2p.options}
\alias{future.p2p.debug}
\alias{R_FUTURE_DEBUG}
\title{Options used by future.p2p}
\description{
Below are the \R options and environment variables that are used by the
\pkg{future.p2p} package.\cr
\cr
\emph{WARNING: Note that the names and the default values of these options may
change in future versions of the package.  Please use with care until
further notice.}
}
\section{Packages must not change these options}{


Just like for other R options, as a package developer you must \emph{not} change
any of the below options.  Only the end-user should set these.
If you find yourself having to tweak one of the options, make sure to
undo your changes immediately afterward.
}

\section{Options}{

\describe{
\item{\option{future.p2p.wormhole}:}{(character)
Specifies the absolute path to the \code{wormhole} executable. If not
specified, a default one will be installed.
}
}
}

\section{Options for debugging}{

\describe{
\item{\option{future.p2p.debug}:}{(logical)
If \code{TRUE}, extensive debug messages are generated.
}
}
}

\section{Environment variables that set R options}{

All of the above \R \option{future.p2p.*} options can be set by
corresponding environment variable \env{R_FUTURE_P2P_*} \emph{when the
\pkg{future.p2p} package is loaded}.
This means that those environment variables must be set before
the \pkg{future.p2p} package is loaded in order to have an effect.
For example, if \code{R_FUTURE_P2P_DEBUG=true}, then option
\option{future.p2p.debug} is set to \code{TRUE} (logical).
}

\examples{
# See debug messages
options(future.p2p.debug = TRUE)

}
