% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormhole.R
\name{find_wormhole}
\alias{find_wormhole}
\title{Find a Wormhole Executable}
\usage{
find_wormhole()
}
\value{
The absolute path to the \code{wormhole} executable as a character string.
Attribute \code{version-string} comprise the \code{wormhole --version} output,
and attributes \code{name} and \code{version} the parsed version string.
If no executable exists, an error is produced.
}
\description{
Find a Wormhole Executable
}
\details{
Unless R option \code{future.p2p.wormhole} specifies an executable, the default
is to download and install \code{wormhole-william} locally and use that one.
}
\keyword{internal}
