% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoogleDriveP2PFutureBackend-class.R
\name{google_drive_p2p}
\alias{google_drive_p2p}
\alias{availableGoogleDriveP2PWorkers}
\title{google_drive_p2p futures}
\usage{
google_drive_p2p(
  ...,
  workers = availableGoogleDriveP2PWorkers(),
  envir = parent.frame()
)

availableGoogleDriveP2PWorkers()
}
\arguments{
\item{workers}{The number of concurrent workers to use.}

\item{envir}{The \link{environment} from where global objects should be
identified.}

\item{\ldots}{Additional arguments passed to \code{Future()}.}
}
\value{
An object of class \code{GoogleDriveP2PFuture}.

\code{availableGoogleDriveP2PWorkers()} returns the number of registered workers on the P2P cluster.
It will always return at least one worker, which is yourself.
\emph{WARNING: This is currently hardcoded to 10 workers, regardless of the number.}
}
\description{
\emph{WARNING: This function must never be called.
It may only be used with \code{\link[future:plan]{future::plan()}}}
}
\details{
A google_drive_p2p future is an asynchronous multiprocess
future that will be evaluated in a background R session.

The Google Drive P2P future backend relies on Google Drive to
distribute futures among a peer-to-peer (P2P) network of R workers.
Users with a Google Drive account can join the P2P network by
being invited to a shared folder.

Users who wish to share contribute the compute power of their
computer should call \code{\link[=google_drive_p2p]{google_drive_p2p()}}.

Users who wish to take advantage of the compute power of the
P2P network should use \code{plan(google_drive_p2p)}.

The Google Drive folder used for orchestrating futures to be processed,
currently processed, and return results is currently hardcoded to
\verb{~/futureverse/future.p2p/}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
googledrive::drive_auth()

## Futures are pushed to P2P Google Drive and 
## results are collected from there
plan(future.p2p::google_drive_p2p, .init = FALSE)

## Create future
a <- 42
f <- future({ 2 * a })

## Get results
v <- value(f)
print(v)
\dontshow{\}) # examplesIf}
}
\seealso{
Users who wish to share contribute the compute power of their computer
should call \code{\link[=google_drive_p2p_worker]{google_drive_p2p_worker()}}.
}
