% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_cluster.R
\name{host_cluster}
\alias{host_cluster}
\title{Hosts a P2P cluster}
\usage{
host_cluster(
  cluster = p2p_cluster(users),
  users = character(0L),
  name = p2p_name(),
  host = "pipe.pico.sh",
  ssh_args = NULL,
  duration = 14 * 24 * 60 * 60
)
}
\arguments{
\item{cluster}{(character string) The name of the p2p cluster.}

\item{users}{(character vector) Names of Pico users who should have access,
in addition to the owner.
The default is a personal cluster that only you have access to.}

\item{name}{The name of the cluster owner as publicized on the P2P cluster.}

\item{host}{(character string) The hostname serving the pico service.}

\item{ssh_args}{(character vector) Optional SSH arguments.}

\item{duration}{Duration (in seconds) to offer this cluster.}
}
\description{
Hosts a P2P cluster
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Connect to personal P2P cluster, which is automatically launched
host_cluster(users = c("bob", "carol"))
\dontshow{\}) # examplesIf}
}
