% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2p_name.R
\name{p2p_cluster}
\alias{p2p_cluster}
\alias{p2p_name}
\title{Gets the default name of the P2P cluster}
\usage{
p2p_cluster()

p2p_name()
}
\value{
\code{p2p_cluster()} return R option \code{future.p2p.cluster}, if set,
otherwise \code{{pico_name}/mycluster}.

\code{p2p_name()} returns the client or worker name of format
\code{"{username}@{hostname}:{pid}"}.
}
\description{
Gets the default name of the P2P cluster

Gets the default name of the client or worker on the P2P cluster
}
