% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2p_name.R
\name{p2p_cluster}
\alias{p2p_cluster}
\alias{p2p_name}
\title{Gets the default name of the P2P cluster}
\usage{
p2p_cluster()

p2p_name()
}
\value{
\code{p2p_cluster()} return the cluster name, if set, otherwise \code{"p2p"}.

\code{p2p_name()} returns the client or worker name, if set, otherwise
"{username}@{hostname}:{pid}"`.
}
\description{
Gets the default name of the P2P cluster

Gets the default name of the client or worker on the P2P cluster
}
