% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pico_p2p_worker.R
\name{pico_p2p_worker}
\alias{pico_p2p_worker}
\title{Launches a P2P worker and adds it to a P2P cluster}
\usage{
pico_p2p_worker(cluster = "chat", name = p2p_name())
}
\arguments{
\item{cluster}{The p2p cluster to contribute to.}

\item{name}{The name of the worker as publicized on the P2P cluster.}
}
\description{
Launches a P2P worker and adds it to a P2P cluster
}
\section{Sequential, single-core processing by default}{

A P2P worker runs sequentially (\code{plan(sequential)}) and is configured
to with a single CPU core to prevent nested parallelization.
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pico_p2p_worker()
\dontshow{\}) # examplesIf}
}
