% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pico_p2p_worker.R
\name{pico_p2p_worker}
\alias{pico_p2p_worker}
\title{Launches a Pico P2P Worker}
\usage{
pico_p2p_worker(channel = "chat", user = pico_user())
}
\description{
Launches a Pico P2P Worker
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pico_p2p_worker()
\dontshow{\}) # examplesIf}
}
