% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PicoP2PFutureBackend-class.R
\name{PicoP2PFutureBackend}
\alias{PicoP2PFutureBackend}
\title{A Pico P2P future is resolved through Peer-to-Peer (P2P) workers communicating via pico.sh and Wormhole}
\usage{
PicoP2PFutureBackend(
  cluster = p2p_cluster_name(),
  host = "pipe.pico.sh",
  ssh_args = NULL,
  ...
)
}
\arguments{
\item{cluster}{The p2p cluster to connect to.}

\item{host}{(character string) The hostname serving the pico service.}

\item{ssh_args}{(character vector) Optional SSH arguments.}

\item{\ldots}{Additional arguments passed to \code{\link[future:FutureBackend-class]{future::FutureBackend()}}.}
}
\value{
A \code{PicoP2PFutureBackend} object
}
\description{
A Pico P2P future is resolved through Peer-to-Peer (P2P) workers communicating via pico.sh and Wormhole
}
\keyword{internal}
