% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PicoP2PFutureBackend-class.R
\name{PicoP2PFutureBackend}
\alias{PicoP2PFutureBackend}
\title{A Pico P2P future is resolved through a Peer-to-Peer (P2P) workers communicating via pico.sh and Wormhole}
\usage{
PicoP2PFutureBackend(
  workers = availablePicoP2PWorkers(),
  cluster = "chat",
  name = pico_user(),
  ...
)
}
\arguments{
\item{workers}{(optional) The maximum number of workers the
backend may use at any time.}

\item{cluster}{The p2p cluster to connect to.}

\item{name}{The name of the client as publicized on the P2P cluster.
The default name is \verb{\{username\}@\{hostname\}:\{pid\}}.}

\item{\ldots}{Additional arguments passed to \code{\link[future:FutureBackend]{future::FutureBackend()}}.}
}
\value{
A PicoP2PFutureBackend object
}
\description{
A Pico P2P future is resolved through a Peer-to-Peer (P2P) workers communicating via pico.sh and Wormhole
}
\keyword{internal}
