% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{worker}
\alias{worker}
\title{Launches a P2P worker and adds it to a P2P cluster}
\usage{
worker(
  cluster = p2p_cluster_name(host = host, ssh_args = ssh_args),
  host = "pipe.pico.sh",
  ssh_args = NULL,
  duration = 60 * 60,
  sandbox = FALSE
)
}
\arguments{
\item{cluster}{The P2P cluster to contribute to.}

\item{host}{(character string) The hostname serving the pico service.}

\item{ssh_args}{(character vector) Optional SSH arguments.}

\item{duration}{Duration (in seconds) to offer working on futures.}

\item{sandbox}{If TRUE, the future is resolved in R WebAssembly (webR),
using the external \code{rw} tool (1).}
}
\value{
Nothing.
}
\description{
Launches a P2P worker and adds it to a P2P cluster
}
\section{Sequential, single-core processing by default}{

A P2P worker runs sequentially and is configured with a single CPU core
to prevent nested parallelization.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
## Start a P2P cluster worker
future.p2p::worker()
\dontshow{\}) # examplesIf}
}
\references{
\enumerate{
\item rw: CLI for webR with Sandboxing Features, \emph{under development},
\url{https://github.com/HenrikBengtsson/rw}.
}
}
