% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pico_p2p_worker.R
\name{worker}
\alias{worker}
\title{Launches a P2P worker and adds it to a P2P cluster}
\usage{
worker(
  cluster = p2p_cluster(),
  name = p2p_name(),
  host = "pipe.pico.sh",
  ssh_args = NULL,
  duration = 60 * 60
)
}
\arguments{
\item{cluster}{The p2p cluster to contribute to.}

\item{name}{The name of the worker as publicized on the P2P cluster.}

\item{host}{(character string) The hostname serving the pico service.}

\item{ssh_args}{(character vector) Optional SSH arguments.}

\item{duration}{Duration (in seconds) to offer working on futures.}
}
\description{
Launches a P2P worker and adds it to a P2P cluster
}
\section{Sequential, single-core processing by default}{

A P2P worker runs sequentially (\code{plan(sequential)}) and is configured
to with a single CPU core to prevent nested parallelization.
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## Start a P2P cluster worker
future.p2p::worker()
\dontshow{\}) # examplesIf}
}
