% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormhole.R
\name{wormhole_send}
\alias{wormhole_send}
\title{Send and Receive Files via Wormhole}
\usage{
wormhole_send(file, code, ...)
}
\arguments{
\item{code}{(character string) Secret wormhole code.}

\item{...}{Not used.}
}
\description{
Send and Receive Files via Wormhole
}
\seealso{
This function relies on the \url{https://pico.sh} services.
}
